+++ app/src/main/java/org/axonframework/samples/trader/app/api/order/BuyOrderPlacedEvent.java
@@ -22,7 +22,7 @@ import org.axonframework.domain.AggregateIdentifier;
  */
 public class BuyOrderPlacedEvent extends AbstractOrderPlacedEvent {
 
-    public BuyOrderPlacedEvent(AggregateIdentifier orderId, long tradeCount, int itemPrice, AggregateIdentifier portfolioId) {
+    public BuyOrderPlacedEvent(AggregateIdentifier orderId, long tradeCount, long itemPrice, AggregateIdentifier portfolioId) {
         super(orderId, tradeCount, itemPrice, portfolioId);
     }
